fix(Service): tighten dependency normalization
Tighten the dependency normalization in the
Service decorator to 100% conform to the
expected overloads.
This fixes some issues in the prior implementation.

Furthermore, the types for the  single-parameter
use-case have been tightened up to ensure the
dependencies key is present.
See 5ca10c4067600daf8751632d032d6d0e5329e8e9.