fix(ContainerInstance): make disposed readonly
This prevents the container being brought "back alive"
after a disposal, which would introduce undefined
behaviour.

One side-effect of this is that in the dispose method,
we have to forcibly cast `this` to `any` to modify the
property.
However, the main focus of this addition is ensuring
consumers of the API will not accidentally change the
value of the property.

Inside ContainerInstance, the property is only changed in one
place, which makes this easier to implement.