fix(ContainerInstance): allow ".on" when disposed
Change the ContainerInstance behaviour
to allow the use of `.of` even when the
container has been disposed.

This means that if the default container
has been destroyed, consumers are still
able to use `Container.of` to create new
containers.

The use of "disposed" was not appropriate
in this context, as the method does not
rely on any instance-specific metadata.