feat(Container): add first revision of Visitor API
Add the first revision of the Visitor API to the container.
The API allows consumers to listen to certain
operations performed in the context of a container.
The implementation is decoupled from that of Containers.

Visitors are a play on the formal concept of the visitor pattern,
which is explained in abstract via the link below:
<https://en.wikipedia.org/wiki/Visitor_pattern>
Some inspiration was also taken from Dagger's SPI types.

This opens up the possibility of extending containers,
including the default, with custom functionality such as
reporting, diagnostics, error recovery and more.

Currently, the published API is deliberately minimal;
it implements a very small subset of functionality.
A good idea is to move slowly on this instead of rushing
the API design and regretting it later on.

As an example of the minimalistic design, some aspects
which were deliberately elided from this rendition include:
  - Allowing visitors to implement custom providers
Such functionality may not be appropriate in the context
of the Container, as it would introduce additional complexity
and a general uncertanty as to the location of metadata.

The API _definitely_ needs some form of revision.
It is not ready to be merged as-is.
No tests have been created yet due to the API most likely changing.
However, a basic structure has been put in place.