fix(ContainerInstance): make .has read recursive correctly
Make the .has method read the recursive option correctly.
Now, if it is set to false, it will correctly claim the symbol
does not exist, even if it is found on the parent.

One optimisation regarding this may be updating the
getIdentifierLocation method to also accept a recursive
parameter and, if that is set to false, do not scan the
parent at all.