feat(website): update theme
Update the code-block theme to throw
when a title is used and it is not terminated.

Also, add a component that allows for
easier referencing of GitHub items
(issues, PRs, etc.)

Finally, add a WorkInProgress component
that signals that a page is being worked on.
Ironically, this component is a WIP too.