test(ContainerInstance): fix tests to use correct reset strategy
Through adding additional testing, I uncovered that the previous
test suite for Service was incorrect, and used the wrong reset strategy.

This meant that the container was not being wiped before each
new test, tainting the environment.  Ultimately, this was uncovered when
testing the interface surrounding the usage of the same ID for multiple
service classes.