feat(website): update hero with new styling
Add a much nicer set of styles to the site.
These styles change the colour of the site,
and add new styles that can be used by
components later on.

Furthermore, the light theme is updated
to somewhat match the dark theme.
While it still looks pretty bad, it's a few
steps closer to normality.