feat(Service): introduce Service type-checking
While this implementation is very... verbose,
it does work and I see no practical way around it
in the current version of TypeScript.

I've tried to condense the documentation by way
of TSDoc inheritDoc tags, but they don't actually
display in solutions such as VSCode.
This makes the documentation worse, as the docs
for each static overload wouldn't be visible without
scrolling to the generic fall-through overload.

Currently, the "everything is an option object" overload
isn't type-checked.  This makes the code a bit less
verbose, but it may be desirable.