refactor(Service): replace ServiceSubject with ServiceWithDependencies
Furthermore, simplify the type-resolution algorithm and remove its
dependency on arrays with a fixed number of items.
The new implementation is compatible with arrays of any size.

A major advantage of this implementation is that it **dramatically**
simplifies the implementation of dependency type-checking,
while also removing the need for multiple Service / JSService overloads.
Overall, this means much less code duplication.