fix(ContainerInstance): preserve non-reconstructable values on import
Prevent ContainerInstance from clearing the value of an imported
service when the service cannot be reconstructed.

A service is classed as "non-reconstructable" when it does not have
a "factory" or a "type" to allow recreation of the value.
This would happen in the case of setValue calls, for example.

This change preserves compatibility with the current API while also
fixing a core bug in ContainerInstance.