feat(ContainerInstance): add creation options
These are a set of options which tell TypeDI
how to handle conflicting containers.
For instance, if a child container is created with
the ID "my-container", and then an ofChild
call is made with the same ID, the pre-existing
container will be returned.

This has a lot of downsides, many of which
are documented in the create options file.

(In fact, I'd almost call this a fix instead of feat.)

The new API doesn't interfere with TypeDI's
default behaviour of allowing conflicts.
Therefore, this isn't a breaking change.