fix(Service)!: re-type service options to comply with spec
BREAKING CHANGE: This will most likely impact consumers of the library who rely
on the structure of the ServiceOptions type.

The ServiceOptions type has been re-typed
to make it function correctly, as according to
the specification both here and upstream.
It now requires either factory, value or type
to be present in the list of service options.

The new type still enforces the optional
inclusion of dependencies, encoded as
an array of service dependencies (which
haven't been type-wrapped yet.)