feat(website): add NPMInstallBlock
This component makes it easy to document how
to install a certain package for multiple package
managers (e.g. NPM, PNPM, Yarn) while also making
it easier to add support for popular PMs later on.

This idea was copied from Docusaurus, which makes
use of a Remark plugin to achieve what we've done
with a much simpler MDX component.