fix(container)!: pass dependencies to factory
Pass parameters declared by services to
the corresponding factory function, if present.
When this happens, built-ins are not guarded.
Instead, they are passed directly to the
factory.  For instance:

```ts
@Service(
  [Number],
  { factory: (number: typeof Number) => instanceOfService }
)
class MyService {}
```

An implementation note worthy of mention is that
currently, the values passed to the factory are
not clear via types.  This will be fixed when
dependency type-checking is implemented.