refactor(*): check dependencies in resolvable wrapper
BREAKING CHANGE: Check each dependency in the
wrapDependencyAsResolvable function.  This is primarily a
performance optimisation, as it means we don't have to
iterate each service' list of dependencies twice.

However, this may break some use-cases.
As Container.set now does the type-checking, any
manually-injected services which do not conform to the
expected set of rules will result in an error being thrown.

One advantage of moving the logic into Container.set, aside
from the above, is that we essentially unify our dependency
handling logic.  Instead of subtly having separate rules for
Service and manually-set services, this ensures that all
services in the container conform to the expected specification.

Additional TSDoc has been added to the Container.set method
to explain the above.