fix(container)!: improve handling of singletons
BREAKING CHANGE

Explicitly avoid caching singletons in containers,
as their value may change and we'd want to ensure
that the consumer gets the most up-to-date value
on each invocation.

This is a breaking change because the prior behaviour
was to cache the value in the container and, if it
changed, any further .get calls for that token would
return an out-of-date value.

Ultimately, this unifies the management of singletons
and makes it easier to keep individual containers in sync.