refactor(npm): remove ESM2015 build
This build has not been actively maintained.

After much research, it seems to be a holdover
from Angular, which used it to allow package
authors to specify a patch to ESM code.
Therefore, the purpose of this has now been
superseded by the "module" field.

This is solidified by the fact that zone.js,
one of Angular's libraries, sets the "es2015"
field to the same value as that of "module"[0].

[0]: <https://github.com/angular/angular/blob/c3349fb2a1b66c29107c8344991716f4e461e95a/packages/zone.js/package.json#L7>

Further reading:
<https://docs.google.com/document/d/1CZC2rcpxffTDfRDs6p1cfbmKNLA6x5O-NtkJglDaBVs/edit#heading=h.k0mh3o8u5hx>