feat(container): introduce resetValuesLazily
This commit introduces a new method to the
ContainerInstance class: resetValuesLazily.
This method returns an iterator which lazily
disposes each service in the container.

The main advantage of this method is that,
for each iteration, it yields both the ID of
the service being disposed, alongside a "status"
member, which signifies the progress of the
service's disposal routine.