fix(transientref): handle constraints correctly
This fixes a case wherein SkipSelf was used, and the TransientRef
constructor was attempting to lookup the identifier's metadata on
the passed-in container, instead of that container's parent.

We have also ensured the create and createOrNull methods perform
correctly, by taking the provided constraints into account.
This commit refactors them to use the new "resolveConstrainedIdentifier"
method of ContainerInstance, which allows us to emulate native
constraint handling instead of having to re-implement it inline.