refactor(container): remove scope from many metadata
Remove the "scope" member from metadata for "many" services.
This simplifies the implementation, as it means we don't have
to set or check for it in the get / set methods.

Ultimately, the scope field makes little sense, as each .set
call results in a different masked token being created.
That masked token is created with the scope passed to the .set
method; that scope may differ across calls for the same group.

Therefore, to elide any confusion around scoping, the member
has been completely removed from the interface.

This commit also removes the useless scope check in the getMany
method.  If an object with the "scope" field set to "singleton"
is passed to .set, the call is passed to the default container.
Therefore, checking whether a group's scope is "singleton" is
pointless, as it will never occur unless the object has been
tampered with (and we don't support that).

This does not affect the existing handling of groups.

Implementation note: Even though we could drop the metadata
interface altogether, it's being kept to allow for expansion
in future.