feat(container): implement ID presence scopes
It's currently not possible to check whether an
identifier is part of a group, or whether it's a
singular item.

This means that, even if Container.has(A) = true,
there is still a possibility that Container.get(A)
will throw.

This makes the API slightly confusing to deal with,
as you have to know which keys form parts of a group,
and which keys don't *before* using the API.