feat(tests): allow imports from bundle files
This commit introduces new functionality to the
testing suite, which allows it to also test the
bundled files, generated by Rollup / Terser.

This helps to ensure that the bundle files are
correct, which is a strategy also employed by
packages such as React.

As of current, this functionality is experimental
and does not work.