feat(core): add web build entry point generator
Add a script that generates a barrel entry point for
the web build.  This allows us to pull-in contrib and
other modules at build-time, as opposed to shipping
them in separate modules.

Also included is a new tsconfig that allows for the
referencing of contributory modules without the use
of a relative path -- this makes it easier to write
the build script without breaking stuff.