feat(scripts): move Rollup cfg to scripts/, refactor
The Rollup configuration file has been refactored so it
is able to create web builds of the Container, which
include all contributory modules with valid entry-points.

The new configuration file has been refactored to use
a more FP-style of programming, which has helped to
reduce the amount of code duplication needed.

The web entry-point generator utility has also been
refactored to make it easier to use.  Asynchronous
functions have been adopted in the hopes of allowing
for a faster build time, though further optimisations
could probably be made here.

The TypeScript configuration also had to be changed
to allow for the migration of the rollup.config.mjs
file to its TypeScript equivalent.  The module
options had to be changed to Node. This doesn't
seem to affect anything else.